* refactor(migrate): rename migrate.ts to utils.ts * feat(utils.ts): add useCollectionReady function to fetch ready collections for migration
