+++ components/migrate/landing/header/Index.vue
@@ -125,7 +125,7 @@ import {
   NeoDropdownItem,
   NeoIcon,
 } from '@kodadot1/brick'
-import useMigrate from '@/components/migrate/migrate'
+import useMigrate from '@/components/migrate/utils'
 
 const { source, sourceSelected, destination, destinationSelected } =
   useMigrate()
